GH-789 Move constants into EdaRegionConnectorMetadata
