--- region-connectors/region-connector-at-eda/src/main/java/energy/eddie/regionconnector/at/eda/utils/DateTimeConstants.java
@@ -1,11 +1,0 @@
-package energy.eddie.regionconnector.at.eda.utils;
-
-import java.time.ZoneId;
-
-public class DateTimeConstants {
-
-    public static final ZoneId AT_ZONE_ID = ZoneId.of("Europe/Vienna");
-
-    private DateTimeConstants() {
-    }
-}
