+++ region-connectors/region-connector-at-eda/src/test/java/energy/eddie/regionconnector/at/eda/services/PermissionRequestFulfillmentServiceTest.java
@@ -25,7 +25,7 @@ import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
 import java.util.List;
 
-import static energy.eddie.regionconnector.at.eda.utils.DateTimeConstants.AT_ZONE_ID;
+import static energy.eddie.regionconnector.at.eda.EdaRegionConnectorMetadata.AT_ZONE_ID;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
